combine functions into fewer files
