--- ppid.c
@@ -1,13 +1,0 @@
-#include "shell.h"
-/**
- * get_ppid - function to return the process ID of the
- * parent process of the calling process
- * Return: ppid
- */
-unsigned int get_ppid(void)
-{
-	unsigned int ppid;
-
-	ppid = getppid();
-	return (ppid);
-}
